+++ packages/common/integration/tests/ivy/helpers/test-extending.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-classes-per-file */
 import { ApplicationRef, Component, Directive, Injectable, NgZone } from '@angular/core';
 
 import { InjectFeatureTestService, InjectNgZone, InjectTestService } from './test-decorators';
@@ -6,22 +7,23 @@ import { FeatureTestService, TestService } from './test-default';
 @Directive({
     selector: '[test-stairs-a]'
 })
-export class TestStairsA {
+export class TestStairsADirective {
     @InjectNgZone()
     public ngZone!: NgZone;
 }
 
 @Injectable()
-export class TestStairsB extends TestStairsA {
+export class TestStairsB extends TestStairsADirective {
     @InjectTestService()
     public testService!: TestService;
 }
 
+// eslint-disable-next-line @angular-eslint/prefer-on-push-component-change-detection
 @Component({
     selector: 'test-stairs-c',
     template: '{{ ngZone.constructor.name }} {{ testService.testField }} {{ featureTestService.constructor.name }}'
 })
-export class TestStairsC extends TestStairsB {
+export class TestStairsComponent extends TestStairsB {
     @InjectFeatureTestService()
     public featureTestService!: FeatureTestService;
 }
@@ -38,6 +40,7 @@ export abstract class AbstractSuperTestDirective {
     protected constructor(public appRef: ApplicationRef) {}
 }
 
+// eslint-disable-next-line @angular-eslint/prefer-on-push-component-change-detection
 @Component({
     selector: 'extending-test-component',
     template: ''